+++ library/src/main/java/com/g2pdev/smartrate/logic/StoreLinkResolver.kt
@@ -16,7 +16,8 @@ internal class StoreLinkResolverImpl : StoreLinkResolver {
             Store.AMAZON -> getAmazonLink(packageName)
             Store.XIAOMI -> getXiaomiLink(packageName)
             Store.SAMSUNG -> getSamsungLink(packageName)
-//            Store.APTOIDE -> getAptoideLink(packageName)
+            Store.APP_GALLERY -> getAppGalleryLink(packageName)
+            //            Store.APTOIDE -> getAptoideLink(packageName) // TODO
         }
     }
 
@@ -60,26 +61,34 @@ internal class StoreLinkResolverImpl : StoreLinkResolver {
         return Single.just(
             StoreLink(
                 link = samsungLinkFormat.format(packageName),
-                alternateLink = xiaomiAlternateLinkFormat.format(packageName)
+                alternateLink = samsungAlternateLinkFormat.format(packageName)
+            )
+        )
+    }
+
+    private fun getAppGalleryLink(packageName: String): Single<StoreLink> {
+        return Single.just(
+            StoreLink(
+                link = appGalleryLinkFormat.format(packageName),
+                alternateLink = appGalleryAlternateLinkFormat
             )
         )
     }
 
     private companion object {
         private const val googlePlayLinkFormat = "market://details?id=%s"
-        private const val googlePlayAlternateLinkFormat =
-            "https://play.google.com/store/apps/details?id=%s"
+        private const val googlePlayAlternateLinkFormat = "https://play.google.com/store/apps/details?id=%s"
 
         private const val amazonLinkFormat = "amzn://apps/android?p=%s"
-        private const val amazonAlternateLinkFormat =
-            "https://www.amazon.com/gp/mas/dl/android?p=%s"
+        private const val amazonAlternateLinkFormat = "https://www.amazon.com/gp/mas/dl/android?p=%s"
 
-        // TODO: Validate these links
-        private const val xiaomiLinkFormat = "market://details?id=%s"
-        private const val xiaomiAlternateLinkFormat =
-            "https://play.google.com/store/apps/details?id=%s"
+        private const val xiaomiLinkFormat = "mimarket://details?id=%s"
+        private const val xiaomiAlternateLinkFormat = "https://play.google.com/store/apps/details?id=%s"
 
         private const val samsungLinkFormat = "https://galaxystore.samsung.com/detail/%s"
         private const val samsungAlternateLinkFormat = "https://galaxystore.samsung.com/detail/%s"
+
+        private const val appGalleryLinkFormat = "appmarket://details?id=%s"
+        private const val appGalleryAlternateLinkFormat = "https://appgallery8.huawei.com/"
     }
 }